requirements.txt
absl-py==1.3.0
accelerate==0.21.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
apex==0.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
audioread==3.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.9
Brotli==1.0.9
cachetools==5.2.0
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.3
cloudpickle==2.2.0
cmake==3.24.1.1
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
confection==0.0.3
contourpy==1.0.6
cuda-python @ file:///rapids/cuda_python-11.7.0%2B0.g95a2041.dirty-cp38-cp38-linux_x86_64.whl
cudf @ file:///rapids/cudf-22.10.0a0%2B316.gad1ba132d2.dirty-cp38-cp38-linux_x86_64.whl
cugraph @ file:///rapids/cugraph-22.10.0a0%2B113.g6bbdadf8.dirty-cp38-cp38-linux_x86_64.whl
cuml @ file:///rapids/cuml-22.10.0a0%2B56.g3a8dea659.dirty-cp38-cp38-linux_x86_64.whl
cupy-cuda118 @ file:///rapids/cupy_cuda118-11.0.0-cp38-cp38-linux_x86_64.whl
cycler==0.11.0
cymem==2.0.7
Cython==0.29.32
dask @ file:///rapids/dask-2022.9.2-py3-none-any.whl
dask-cuda @ file:///rapids/dask_cuda-22.10.0a0%2B23.g62a1ee8-py3-none-any.whl
dask-cudf @ file:///rapids/dask_cudf-22.10.0a0%2B316.gad1ba132d2.dirty-py3-none-any.whl
datasets==2.13.1
dbus-python==1.2.16
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.6.0
diffusers==0.18.2
dill==0.3.6
distributed @ file:///rapids/distributed-2022.9.2-py3-none-any.whl
distro==1.4.0
einops==0.6.1
entrypoints==0.4
et-xmlfile==1.1.0
evaluate==0.4.0
exceptiongroup==1.0.4
execnet==1.9.0
executing==1.2.0
expecttest==0.1.3
fairscale==0.4.8
faiss-gpu==1.7.2
fastjsonschema==2.16.2
fastrlock==0.8.1
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.38.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
google-auth==2.15.0
google-auth-oauthlib==0.4.6
graphsurgeon @ file:///workspace/TensorRT-8.5.1.7/graphsurgeon/graphsurgeon-0.4.6-py2.py3-none-any.whl
grpcio==1.51.1
h5py==3.9.0
HeapDict==1.0.1
huggingface-hub==0.16.4
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
hypothesis==5.35.1
idna==3.4
imageio==2.9.0
importlib-metadata==5.1.0
importlib-resources==5.10.1
inflate64==0.3.1
iniconfig==1.1.1
intel-openmp==2021.4.0
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.17.3
jupyter-tensorboard @ git+https://github.com/cliffwoolley/jupyter_tensorboard.git@ffa7e26138b82549453306e06b535a9ac36db17a
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyterlab==2.3.2
jupyterlab-pygments==0.2.2
jupyterlab-server==1.2.0
jupyterlab-widgets==3.0.8
jupytext==1.14.4
kiwisolver==1.4.4
langcodes==3.3.0
librosa==0.9.2
lit==16.0.6
llvmlite==0.39.1
locket==1.0.0
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mediapipe==0.10.2
mistune==2.0.4
mkl==2021.1.1
mkl-devel==2021.1.1
mkl-include==2021.1.1
mock==4.0.3
mongoengine==0.22.1
mpmath==1.2.1
msgpack==1.0.4
multidict==6.0.4
multiprocess==0.70.14
multivolumefile==0.2.3
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.4.10
numba==0.56.4
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-dali-cuda110==1.20.0
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvidia-pyindex==1.0.9
nvtx==0.2.5
oauthlib==3.2.2
odsm==1.0.52
omegaconf==2.3.0
onnx @ file:///opt/pytorch/pytorch/third_party/onnx
open-clip-torch==2.20.0
openai==0.27.8
opencv @ file:///opencv-4.6.0/modules/python/package
opencv-contrib-python==4.8.0.74
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==23.1
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
pigeonXT-jupyter==0.6.1
Pillow==8.1.2
pkgutil_resolve_name==1.3.10
platformdirs==2.6.0
pluggy==1.0.0
polygraphy==0.43.1
pooch==1.6.0
preshed==3.0.8
prettytable==3.5.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.19.0
pyarrow @ file:///rapids/pyarrow-9.0.0-cp38-cp38-linux_x86_64.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybcj==1.0.1
pybind11==2.10.1
pycocotools @ git+https://github.com/nvidia/cocoapi.git@8b8fd68576675c3ee77402e61672d65a7d826ddf#subdirectory=PythonAPI
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.7.4
Pygments==2.13.0
PyGObject==3.36.0
pylibcugraph @ file:///rapids/pylibcugraph-22.10.0a0%2B113.g6bbdadf8.dirty-cp38-cp38-linux_x86_64.whl
pylibraft @ file:///rapids/pylibraft-22.10.0a0%2B81.g08abc72.dirty-cp38-cp38-linux_x86_64.whl
pymongo==3.11.2
pynvml==11.4.1
pyparsing==3.0.9
pyppmd==0.18.3
pyre-extensions==0.0.29
pyrsistent==0.19.2
pytest==7.2.0
pytest-rerunfailures==10.3
pytest-shard==0.1.2
pytest-xdist==3.1.0
python-dateutil==2.8.2
python-hostlist==1.22
pytorch-quantization==2.1.2
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==24.0.1
pyzstd==0.15.9
raft-dask @ file:///rapids/raft_dask-22.10.0a0%2B81.g08abc72.dirty-cp38-cp38-linux_x86_64.whl
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
responses==0.18.0
rich==9.6.1
rmm @ file:///rapids/rmm-22.10.0a0%2B38.ge043158.dirty-cp38-cp38-linux_x86_64.whl
rsa==4.9
safetensors==0.3.1
scikit-image==0.19.3
scikit-learn @ file:///rapids/scikit_learn-0.24.2-cp38-cp38-manylinux2010_x86_64.whl
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.11.0
soupsieve==2.3.2.post1
spacy==3.4.4
spacy-legacy==3.0.10
spacy-loggers==1.0.4
sphinx-glpi-theme==0.3
srsly==2.4.5
ssh-import-id==5.10
stack-data==0.6.2
starlette==0.14.2
sympy==1.11.1
tbb==2021.7.1
tblib==1.7.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorrt @ file:///workspace/TensorRT-8.5.1.7/python/tensorrt-8.5.1.7-cp38-none-linux_x86_64.whl
terminado==0.17.1
texttable==1.6.7
thinc==8.1.5
threadpoolctl==3.1.0
tifffile==2023.7.10
timm==0.6.7
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torch-tensorrt @ file:///opt/pytorch/torch_tensorrt/py/dist/torch_tensorrt-1.3.0a0-cp38-cp38-linux_x86_64.whl
torchaudio==2.0.2
torchtext @ git+https://github.com/pytorch/text@fae8e8cabf7adcbbc2f09c0520216288fd53f33b
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.7.1
transformers @ git+https://github.com/huggingface/transformers@3ec10e6c76362191b61260300fe1d6173a8dd7e1
treelite @ file:///rapids/treelite-2.4.0-py3-none-manylinux2014_x86_64.whl
treelite-runtime @ file:///rapids/treelite_runtime-2.4.0-py3-none-manylinux2014_x86_64.whl
triton==2.0.0
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.7.1
ucx-py @ file:///rapids/ucx_py-0.27.0a0%2B29.ge9e81f8-cp38-cp38-linux_x86_64.whl
uff @ file:///workspace/TensorRT-8.5.1.7/uff/uff-0.6.9-py2.py3-none-any.whl
urllib3==2.0.3
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.8
xarray==2022.6.0
xdoctest==1.0.2
xformers==0.0.20
xgboost @ file:///rapids/xgboost-1.6.2-cp38-cp38-linux_x86_64.whl
xxhash==3.2.0
yarl==1.9.2
zict==2.2.0
zipp==3.11.0
